requirements.txt
openpyxl==3.1.2
pandas==2.2.2
python-slugify==8.0.4
seaborn==0.13.2
tqdm==4.66.2
